test_reporting/jupyter-notebooks/RebootDataIngestor/requirements.txt
azure-kusto-data
jinja2
